Full re-format on 1.12.2 backend + 1.12.2 backend in progress

+++ shared/main/java/net/ccbluex/liquidbounce/api/minecraft/network/play/server/ISPacketEntityTeleport.kt
@@ -6,7 +6,7 @@ interface ISPacketEntityTeleport : IPacket
 {
 	val entityId: Int
 
-	val x: Int
-	val y: Int
-	val z: Int
+	val x: Double
+	val y: Double
+	val z: Double
 }
